Renamed dir and added login screen
