+++ app/src/main/AndroidManifest.xml
@@ -14,13 +14,17 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
         <activity
-            android:name="io.helpful.android.HomeActivity"
+            android:name="io.helpful.android.ui.LoginActivity"
             android:label="@string/app_name" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+        </activity>
+        <activity
+            android:name="io.helpful.android.ui.HomeActivity"
+            android:label="@string/title_activity_login" >
         </activity>
     </application>
