+++ app/src/main/java/io/helpful/android/ui/LoginActivity.java
@@ -1,4 +1,4 @@
-package io.helpful.android;
+package io.helpful.android.ui;
 
 import android.app.Activity;
 import android.app.ActionBar;
@@ -11,12 +11,14 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.os.Build;
 
-public class HomeActivity extends Activity {
+import io.helpful.android.R;
+
+public class LoginActivity extends Activity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_home);
+        setContentView(R.layout.activity_login);
 
         if (savedInstanceState == null) {
             getFragmentManager().beginTransaction()
@@ -25,20 +27,14 @@ public class HomeActivity extends Activity {
         }
     }
 
-
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
-        
-        // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.home, menu);
+        getMenuInflater().inflate(R.menu.login, menu);
         return true;
     }
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
-        // Handle action bar item clicks here. The action bar will
-        // automatically handle clicks on the Home/Up button, so long
-        // as you specify a parent activity in AndroidManifest.xml.
         switch (item.getItemId()) {
             case R.id.action_settings:
                 return true;
@@ -46,9 +42,6 @@ public class HomeActivity extends Activity {
         return super.onOptionsItemSelected(item);
     }
 
-    /**
-     * A placeholder fragment containing a simple view.
-     */
     public static class PlaceholderFragment extends Fragment {
 
         public PlaceholderFragment() {
@@ -57,7 +50,7 @@ public class HomeActivity extends Activity {
         @Override
         public View onCreateView(LayoutInflater inflater, ViewGroup container,
                 Bundle savedInstanceState) {
-            View rootView = inflater.inflate(R.layout.fragment_home, container, false);
+            View rootView = inflater.inflate(R.layout.fragment_login, container, false);
             return rootView;
         }
     }
