+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':main'
+include 'app'
